Fixed damage scaling for real this time

+++ Items/Weapons/Keybrand.cs
@@ -37,7 +37,7 @@ namespace KeybrandsPlus.Items.Weapons
         }
         public override void ModifyHitNPC(Player player, NPC target, ref int damage, ref float knockBack, ref bool crit)
         {
-            float t = target.life / target.lifeMax;
+            float t = (float)target.life / (float)target.lifeMax;
             float lerpValue = Helpers.KeyUtils.GetLerpValue(1f, 0.1f, t, true);
             float damageBoost = 1.5f * lerpValue;
             damage = (int)(damage * (1 + damageBoost));
