+++ Items/Weapons/TrueKeybrandD.cs
@@ -40,7 +40,7 @@ namespace KeybrandsPlus.Items.Weapons
         }
         public override void ModifyHitNPC(Player player, NPC target, ref int damage, ref float knockBack, ref bool crit)
         {
-            float t = target.life / target.lifeMax;
+            float t = (float)target.life / (float)target.lifeMax;
             float lerpValue = Helpers.KeyUtils.GetLerpValue(1f, 0.1f, t, true);
             float damageBoost = 2f * lerpValue;
             damage = (int)(damage * (1 + damageBoost));
